Replace get/set methods with passed-in functions
I didn't like the weirdness of get/set having to take the first, curried
arg, then call the stored _get/_set values with that arg, as it didn't
match up between get, which took the structure, and set, which took the
value to set, and returned a partially applied function, which would
take the structure. This was hard to see in the code, and I realized I
just don't need to forward things through those methods. I just need to
store get/set as properties. Well, I hope that's all I need...