Add ImpureLens
I don't know if this is a great idea, but I ran into two issues trying
to use this library in Autodesk Maya, to create a lens to target the xyx
translations of transforms. The first was that, because I was simply
calling impure "actions" on objects in the scene, and not manipulating
values passed into a function, which I could pull apart to manipulate
some internal element, and then reassemble into a copy, immutably, I was
forgetting to return the new value. The second was that, when I
remembered to return the value, which ballooned the single call to the
command in Maya, I further realized that I could lock that value in
Maya, which would cause set and modify to fail silently without changing
it, and my cumbersome functions were just returning the value passed in,
as though everything had worked out, which is a bad assumption to make
when sending orders into the world outside the program. ImpureLens gets
around this by simply deriving (by subclassing) everything from Lens,
but overriding the init to wrap the passed-in setter, such that it also
calls the passed-in getter again, after the set operation, and returns
that new (or still old, if things failed) value.